Indexing of Facebook Ads
It is now possible to configure indexes to group Facebook ads (e.g a
common usage would be to define an index on "Constituency".

A central "DB" class is repsonsible for storing all of the ads, and
notifying each indx whenever a new ad is added / an existing add is
updated.