Extract a numeric ID for each Ad
Facebook ads now have a numeric ID that can be index on - so that the
keytype no longer requires comparison of std::strings to sort / access
ads.

We extract the id from the page url. Facebook's docs say the link itself
is a valid key - but since the rest of the link simple a compbination
of:
   - Boiler plate address (https://facebook etc)
   - Our API token (same for every add)
We strip the rest, and just use the id in the link.