Serialization / Deserialization Indexes to disk
It is expensive to re-index the full set of ads from scatch each time -
and largely pointless. Each time we have to re-compute the upper-case
variants of the seach strings + do potentially 1,000s of string comparissions
we've already done on a previous parse.

This change allows the indexer to secure the state of its index to disk
when it shutdown down, and restore it before processing the Facebook ad
dump. This means that only new ads go through a full re-index. Existing
ads will be updated to have their new impression / spend / demographic
values - but we do not have to do all the upper-case calcuation / string
searches for them>